refactor adopters into its own app
